+++ src/main/java/com/arangodb/ArangoCollection.java
@@ -54,7 +54,7 @@ import com.arangodb.model.SkiplistIndexOptions;
  * @see <a href="https://docs.arangodb.com/current/HTTP/Collection/">Documents API Documentation</a>
  * @author Mark Vollmary
  */
-public interface ArangoCollection {
+public interface ArangoCollection extends ArangoSerializationAccessor {
 
 	/**
 	 * The the handler of the database the collection is within