+++ src/main/java/com/arangodb/ArangoDatabase.java
@@ -56,7 +56,7 @@ import com.arangodb.model.TraversalOptions;
  * @see <a href="https://docs.arangodb.com/current/HTTP/AqlQuery/">Query API Documentation</a>
  * @author Mark Vollmary
  */
-public interface ArangoDatabase {
+public interface ArangoDatabase extends ArangoSerializationAccessor {
 
 	/**
 	 * Return the main entry point for the ArangoDB driver