+++ src/main/java/com/arangodb/ArangoEdgeCollection.java
@@ -34,7 +34,7 @@ import com.arangodb.model.EdgeUpdateOptions;
  * @see <a href="https://docs.arangodb.com/current/HTTP/Gharial/Edges.html">API Documentation</a>
  * @author Mark Vollmary
  */
-public interface ArangoEdgeCollection {
+public interface ArangoEdgeCollection extends ArangoSerializationAccessor {
 
 	/**
 	 * The the handler of the named graph the edge collection is within